requirements.txt
exceptiongroup==1.2.0
iniconfig==2.0.0
packaging==23.2
paho-mqtt==2.0.0
pluggy==1.4.0
pytest==8.0.2
python-dotenv==1.0.1
tomli==2.0.1
